+++ spec/mongoid/extensions/true_class_spec.rb
@@ -2,6 +2,13 @@ require "spec_helper"
 
 describe Mongoid::Extensions::TrueClass do
 
+  describe "#__sortable__" do
+
+    it "returns 1" do
+      true.__sortable__.should eq(1)
+    end
+  end
+
   describe "#is_a?" do
 
     context "when provided a Boolean" do