+++ spec/mongoid/relations/embedded/many_spec.rb
@@ -3288,4 +3288,23 @@ describe Mongoid::Relations::Embedded::Many do
       end
     end
   end
+
+  context "when the embedded relation sorts on a boolean" do
+
+    let(:circuit) do
+      Circuit.create
+    end
+
+    let!(:bus_one) do
+      circuit.buses.create(saturday: true)
+    end
+
+    let!(:bus_two) do
+      circuit.buses.create(saturday: false)
+    end
+
+    it "orders properly with the boolean" do
+      circuit.reload.buses.should eq([ bus_two, bus_one ])
+    end
+  end
 end