+++ CHANGELOG.md
@@ -795,6 +795,8 @@ For instructions on upgrading to newer versions, visit
 
 ### Resolved Issues
 
+* \#2122 Allow embedded documents to sort on boolean fields.
+
 * \#2119 Allow `Criteria#update_all` to accept atomic ops and normal sets.
 
 * \#2110 `Model#touch` now works properly on embedded documents.